+++ hr-resource/src/lib/mutation-resolvers.ts
@@ -1,23 +1,7 @@
 import { GraphQLError } from "graphql";
-import { Maybe, MutationResolvers, QueryResolvers, Resolver } from "./gql-codegen/graphql.js";
+import { MutationResolvers } from "./gql-codegen/graphql.js";
 import prisma from "./prisma.js";
 import validator from "validator";
-
-export const qResolverCurrentUser: QueryResolvers['currentUser'] = async (
-  _root,
-  _args,
-  { user, organization, roles }
-) => ({
-  ...user,
-  roles: roles.map(role => ({ ...role, hourlyWage: role.hourlyWage.toNumber() })),
-  organization,
-  dateJoined: user.dateJoined.toISOString(),
-  dateOfBirth: user.dateOfBirth.toISOString(),
-});
-
-export const qResolverRoles: QueryResolvers['roles'] = async (_root, _args, { roles }) => await prisma.roles
-  .findMany()
-  .then(roles => roles.map(role => ({ ...role, hourlyWage: role.hourlyWage.toNumber() })));
 
 export const mResolverCreateUser: MutationResolvers['createUser'] = async (
   _root,
@@ -41,7 +25,29 @@ export const mResolverCreateUser: MutationResolvers['createUser'] = async (
     throw new GraphQLError('User with same email already exists', { extensions: { code: 'CONFLICT' } });
   }
 
-  return await prisma.users
+  const x = prisma.$transaction(async (tx) => {
+    const newUser = await prisma.users.create({
+      data: {
+        firstName: input.firstName,
+        middleName: input.middleName,
+        lastName: input.lastName,
+        email: input.email,
+        dateOfBirth: new Date(input.dateOfBirth),
+        dateJoined: new Date(input.dateJoined),
+        city: input.city,
+        country: input.country,
+        phone: input.phone,
+        pincode: input.pincode,
+        province: input.province,
+        streetName: input.streetName,
+        addressL2: input.addressL2,
+        organization: { connect: { id: organization?.id } },
+        UserRoles: { createMany: { data: input.roleIds?.map(roleId => ({ roleId })) ?? [] } },
+      },
+    });
+  })
+
+  const newUser = await prisma.users
     .create({
       data: {
         firstName: input.firstName,
@@ -60,6 +66,7 @@ export const mResolverCreateUser: MutationResolvers['createUser'] = async (
         organization: { connect: { id: organization?.id } },
         UserRoles: { createMany: { data: input.roleIds?.map(roleId => ({ roleId })) ?? [] } },
       },
-    })
-    .then(user => user.id);
+    });
+
+  return newUser.id;
 };
